Major update, grp file structure
An interrim version - content will be moving at least once,
but this is a working model, including the most recent change
to the cgi-parse.py so that it actually works.  Also includes
the new cLdate.py date routines and a partial rb.py - precursor
to rebuild.py.   Next up is one more roll on the interal directory
structure and off we go.