Now schedules MAMP and web browswer
Added clSchedule.py to handle some of the scheduling tasks,
so far, running MAMP Pro and the browser.   Could  also add
scheduling of other browsers, as well as the Interarchy and
other scheduling that is currently done.   Tightened up a
few bits and basically in good shape.

Next up would likely be the addition of a primitive page edit
selector.