OptionLoad - fixed main_frame
Fixed problem with the main_frame setup and the grid method.
Now have colors set by the new .coLab_tkOptions file.  Setup
here on the Mac Air is seriously short of font files but
seems to be working fine.  No colors on th buttons or option
menus at this tie - apparently a bug in Tkinter.