Fixed the part problem
Part changes on edit were not getting picked up - the long
name was getting saved.   Added the proper dict to the structure
and it's working well.

Still seeing it not get set correctly on occasion, but the last
five tries in a row worked just fine.  (May need to try change
the song as well.)