Adds hi-res and "rms" to the graphics
Two birds with one dirt clod of source.  Simply
changing height and width (still globals... for now)
in imagemaker.py got most of it to simply work.   Took
an explict 640x480 out of ffmpeg.sh, and redid a couple
of hard-coded constants.

Still need to rig up a resolution scheme - either many more
resolutions, or the option to set it on each page.  Or both.

This version of the rms encoder still does the odd draw-back
to the rms value when the min/max is inside.  Looks cool - may
stick with it - but in either case, this one has it.