Fix for frame count, prog bar
This changes the number of frames to match the frame rate,
e.g. 2 seconds at 30fps will have 60 images - before it
was 61.  And 61 may be correct.

But - this does fix the progress bar issue...
Fixes #14

Also, the previous fix included fixes for hi-res graphics,
Fixes #2
and the rms feature:
Fixes #13

Though both of these will need work for additional encoding
for various platforms.

Also - a bit of code that calculates the fps has had fractional
fps removed - but I just learned that ffmpeg *can* do fractional
fps - but it needs to be expressed as a fraction.  So that'll need
some work too.  (But:  yea!)