Initial Placeholders for Song edit
Just a duplication of the Page edit methods for the song.  Likely
want to consider combining these - perhaps into a base class.
But this one works to the point of offering the correct list of
songs.   Fields need to be reworked for the song.  Parts might
be an interesting challenge.