Clean up in prep for rework
This is a functional version - major addition is a Menu_row for the
resolution (media_size).   Also removed "song" specific code from the
Menu_row and made it a base class for Song_menu_row.

Changed a_type (named to put it at the top of a debug list) to obj_type.

Adds the whole Sizes class and such in the config module.