Big change: separate renders - canvas for start/end
This is a very large set of changes for the "HackerFest-jk" as well as
the following week.

The major change over the first weekend was to split out
the media and web generation from the main thread.
There is now a render engine that handles the creation
of the given resolution, and all "sub" resolutions.  Also
creates the proper html5 video source tags to pick them
up.

The new "media_size" is a stored string describing the
media dimensions, and the "next" in the chain (will include
other info eventually, related to font and other layout
sizes).

Page objects can now be created blank and loaded as
needed.

Large amounts of code in render.py constitute the
mojor portion of the render engine.

The new clGraphEdit.py module institutes a nearly complete
interface for setting the start and end points in the music
on a screen shot.

Left to do: in corporate the new GraphEdit module into
page create/edit for new screen shots: this will require
keeping track of the original width since we want to do
this on the full sized graphic rather than the scaled version.

Results:
Fixes #18
Mostly fixes #17 - need check pointing