Video scaling / cursor
Mostly this includes code to correct the cursor position based on
image/video size.   Also scales the time boxes to the sqrt of the size
ratio to the base.

Tweaked ffmpeg a bit - seems to be a better balance of size and
quality.   Also took the browser scheduling out of a thread - that
fixed things!