Changes for scaled graphics
A number of changes - including some serious reorganizing to support
scaling the text, lines and such for different resolutions.   Text,
including the time boxes, are scale at a ratio of the square root of
the ratio to the base size - looks good.   Also added a small, one
might say elegant, resolution tag next to the lower right/elapsed time
box.  Went around on that a couple of times - a failed attempt is
currently in imagemaker.make_images_x - simpler to do - but, ahem,
doesn't work: transparency isn't supported that way.

Delayed the destruction of the rendering Toplevel boxes - seems to make
the hangs better - but not yet solved.  And it's a kludge.

Still issues with the way resolution and start/end are handled - these
will probably require a rework of the U/I.

Moved colors to clColors - capitalized the names since they're
constants.

Added some helpful test cases into rebuild - sped development up -
surprise - need to stay with that.