Audio fixup, ffmpeg stdout improvement
Finally nailed the stupid issue with a different number of clips
reported depending on the resolution (wrong var in chunk_offset).
Redid the way the ffmpeg output is handled - more standard  - still
didn't fix the hand between renders.  A few changes around that.
However, working well and I should be able to start on redoing the
menus tomorrow.