Major Menu Fixes
Much better menu stuff - edit page is (well, was) very close - but the
restructuring of the row classes including the  value lookup that was
in the Menu_row.return_value() method and is now in the
Menu_row.handle_menu() row broke creating a new page - the fix for that
may have broken edit.

One notable change in clFunctions:  the "parent" value in the row
classes is now changed to "editor"  it was just getting too confusing.

That said, all the menus for selecting the sound and graphics appear to
be working and  correctly refreshing the screen