Menus are …. Done!
This should complete the basic requirements for issues.   In other
words, this Fixes #19  - the Redo of the menus.

The current code retains the old style buttons, but under them are the
much improved menus- these seem to be working solidly - as far as I can
tell at this point.

Seems to also have a pretty solid render engine and other features.
Clearly will need a new menu interface soon as the single menu list
option is already running close to physical display limits.

Also - discovered more than one file was set up for spaces rather than
tabs…   I'm going to try to "enforce" tabs-4 from here on out.

(This explains most of the changes noted in cltkutils and coLab.

Oh - and I played with shortening the time boxes - seems to work just
fine with a height of 23 vs. the old 35.  A width of 55 seems to work
well with long times (longest so far: 121 minutes)