Several enhancements, dyn. menus, matching, colors
More improvements: top level menus are working well.  Sound and graphic
file menus are now contextual (should export making them "disabled"
instead).  This fixes #19.

Changed the label in GraphEdit to use a cltkutilsPopup - better - still
not there.  Also moved the "Done" button to the middle - was not
visible in some tiny graphics - at least it's moderately usable now in
that case.

Added has_filetype to clutils - this is used to set the context menus
based on whether certain file types are in a dir.

Added a "dark" header for the groups which looks much better on the
main panel.

Moved a few more colors into clColors - need to get the red and green
references in clFunctions switched over too.

clFunctions now checks the size of the graphic *after* the crop step -
this fixes #26.

Oh - and an early fix sometime back put the "aiff" option into the file
browser option in the correct place - this fixes #20 - finally.

Still some issues with the pop-up - not where I want it, the frame is
not getting cycles - may be a way to do that with and "after" and
"update-idletasks".

All in all - coming along nicely.

-jk