Restarting tweaks
Getting back into it - a number of small changes  - some left over from
before Christmas.

Added initial support for utf-8 - mostly in the page and song data
files but also in the main coLab file - appears to allow 8-bit char
data - but not yet handled correctly in either html or TrueType.

Made a chnage in songgen to allow a song to be passed in… if not - the
songlist in the group is used (previously, only the latter was allowed).