Initial unicode support
Handles most of the items related to issue #30 - converts the internal
vars for titles and descriptions to be unicode strings, adds the proper
utf-8 coding comments to the data files and related source files, adds
the <meta> tag to the html files to support utf-8 and…  converts the
utf-8 to ASCII when writing the graphics using unicode - none of the
current fonts seem to support it - but there is more research to be
done.

Currently - there is a bug added related to this - the dict is not
decoding properly - could be a coding issues - *might* be a manual data
structure issue.   (Also dragging along a .gitignore change from the
previous commit.)