More ASCII - UTF-8 cleanup
Rest of the fix for issue #30 - not closing yet  - holding has an
enhancement for finding utf-8 fonts.

Also added a .strip() to the titles and descriptions to keep the extra
blank lines from accumulating on the descriptions, in particular - and
to get rid of leading blank lines on some older pages.