Fix row status
This fixes a LOT of problems with row status.  Changed "set_status" to
"post_status" (and then added the set function back in…  but it's much
tighter now - and takes the editable flag into account.   Added status
fields to a couple of ittems - should consider moving these to the
Data_row parent class.

This seems to be working in pretty much all context  - at least all the
ones I've tested so far.

Fixes #44.