string to filename and friends
A couple of small changes - somewhat in prep for the reworking of the
way we create pages and songs (and presumably groups and projects?)

The big addition are is the pair of functions string_to_filename and
string_to_unique - to turn an arbitrary string into a filename (in this
case, a dir name) that should  be legal on all systems (one can hope)
and in the second case, unique within the passed directory (a counter
is added to the end until the filename is unique).

I plan to branch after this - to isolate the changes to page creation
so I can continue to use the tool while working on the page creation,
which is likely to take a while.