Some fixes - editing mostly
Fixes #38 - needed to do a read on each row before the posting…  had to
add dummy reads to a few.

Also fixes #45 - the tk.messagebox.askquestion doesn't return
true/false, but a string.

Also - probably *doesn't* fix #33 - I think I got that earlier.  But it
does appear to be fixed.