Var set and read fixup
Reworked how the read() and return_value() methods work (could
conceivably be combined - but there may be some flexibility in
retaining the separation for now.

For the data rows in clfunctions, the read() method is assumed to read
the value from the form, while return_value() returns the current value
- generally by doing a a read() first.

Reworked how menus are done - default is now only used to define a
value to display different from "value" (like with the sound and
graphic option menus).

Fixes #38

Fixed the geometry of the adjust window (needs work as per Issue #29).
Also added a geometry option to the cltkutil.Popup class.

Rewrote the help text - IE now works (so far as we know).