Click to Locate
Some major changes to allow the user to click on the video and locate
to that spot.   Seems to be working in "web page mode" on all browsers,
 but only on some in graphical mode.

Implements a new, shared javascript file: video_engine.js which allows
the various events to drive the video.   This does much of the work
needed for Issue #46.

None remaining issue:  still some scaling issues on screen shots in
full screen - full screen doesn't work well on Safari or at all on IE.

Current ffmpeg encoding of webm may also be a problem.