Buttons working
No side buttons yet - but this does the trick and will let me check the
rest.

Renamed the button ID to make it more clear.   Mostly the changes are
in the video_engine.js - mute and seekbars now work, when present,
otherwise, uses buttons.  Mostly it's the new code to set up and handle
the locator buttons.