Font Work
Much improved font work - as per issue #22 - but not quite done.

I rewrote the old fontlib class in clutils as FontLib, having it trace
down the fonts.  Currently being instantiated a LOT which is goofy, but
a decent test for now.   Also  - instead of the file name, should use
the font name as is pulled out in fontimagemaker (which now actually
finds all the fonts and builds the graphic to match!