Intial Timeline work
Convert the top line to a time line - this is the first part of that -
the older top text is moved down / consolidated or tossed.

The top line "upTickText" stuff is working - we just need to space them
out properly.

Also: added support for issue #48 - OpenTypeFonts now work and the
fonts themselves are now (at least) out of git.