Logging Complete
This fixes #41 - should be pretty much all (or nearly so) of the
internal debugging print statements - converted to logging calls.

Also and fixed (I think) a problem with cldate - was calling
convert_vars a second time, during the save.  Had to “fix”
make_dur_string to account for that.

Still a problem with stray png files being created in …/Resources/Fonts
- added some logging for that.