Data Value Bug Fixes
This one should fix #54 as well as fixing #50 (possibly the same bug).
Also fixes #51, which is was simply not implemented - imagemaker.py

The first two of these were apparently caused by using the
imp.load_source module - which conveniently caches values - not what we
want.   Turns out: reading the entire file into a var, then executing
it does what we want;  that change is in the clclasses.import_data
function.

An early commit fixed #48 - but there is no UTF-8 support yet - or
someone's doing it wrong.  That issue is #30 and we'll keep that one
open.

Also substituted the ProximaFontauriItalic.otf font for the
BrentfordBoldItalic.otf font - the latter has issues (in PIL) with
descenders - in particular: 'y'.