Location/Time Tagging Works!
This one should do it - not only do the Javascript <span> tags work -
I've added the
capability to add actual URL options to go to either a time or locator,
and those URLs appear on the Comments page as well as in the sent
email.   This fixes #55.

There seems to be a problem yet with the URL's not always working - I
think the problem is the page is not fully loaded - that problem is
noted in #56.

I'm going to close #46 as well - these changes are somewhat related,
but the majority of the work there is done.

I've migrated the cgi-parse.py script (which doesn't do anything like
what it says anymore) now called html_encode.py and also brought
evalfix.py (similar)  and html_encode.py into the Code dir - and
reworked them both work as a stand-alone and a callable python
functions.