Right Click - Chapter 1
The basics are working - can do a right-click context menu in all
browsers on every platform (haven't tested Linux yet - wcgw?).

Major addition is the ContextJS package, which in turn references
jQuery (v1.11.1) and bootstrap.

So far, just an arbitrary hard-coded menu - but it proves that it
basically works.

Still need to test the code generation - this is mostly just copied
over from the hand-edited index.shtml file.

Also - the official beginning of the end of the clickdiv.

-jk