Major PHP update
This includes a large portion of the php functionality for issue #58
Still needs some tweaking and the locator hidden vars and html5 video
content still need to be generated.   Javascript is not running - most
likely due to the missing bits.

Also includes the initial state of the php include files.