php - 95% ?
This feels like most of the code necessary to close issue #58  - but
not quite.  This brings in the comments, adds the additional geometry
to make ClickIt! work in regular and full screen modes (except for
Apple based browsers), fixes the top graphic, and basically deals with
the out line and video that was causing the top half of the locator
buttons to not work.

Still need to tweak the Creation date and duration bits.

Major addition: clGeo.py - likely a temporary  solution, but a way to
keep the formatting info in one place and still have both the
python-based html generation and php generation tied to the same
references.