php conversion complete
This basically fixes #58 - a couple of small changes left to be dealt
with - the created and update times - need to figure out how make this
work on more than just the latest revs.

Introduces page_refer.php - intended to be dropped on to every php
compatible page (should be almost all).