Rework on images
This implements most of the functionality in #61 - but there are some
problems - esp. wrt to existing pages - largely untried/untested.
Implements versioning and will need to use that to handle things like
creating an initial PageImage.png.   Also  - need to look into creating
the sound file separately from creating the screenshot file.

But much of this is working and it can create a new page and edit it,
last I  checked.

-jk