Image alignment… nearly...
Fixes… (almost) #60 - a one line change in
clFunctions.Graphic_menu_row_soundfile.file_load to force
page.use_soundgraphic to True any time a new (or old) sound file is
loaded vastly simplified things.

Added a new option for the screenshot to simply reuse the last one, and
things flow much smoother…

But…  the posters are not getting generated from the proper base image
- so until then we can consider #60 to be quite done yet.

Also fixes a couple of problems with comments that are outflows of the
php conversion (!desc_title! and explicit index.shtml)