Most of mp3
Issue #61 is almost there… the major pieces are here:
generates mp3 using ffmpeg and lame while its creating the graphics
(recreated for each video - no big deal).  Automatically links with a
"download" button and even (sorta) puts the song name artist in the mp3.

Still bit klunky, and it may be time to consider moving the scheduling
of mp3 into the python code where we have better access to vars, etc.

Need to finally do something with the CSS and buttons - make the
download and soundinfo button look more like the locators… at least for
now.