Adds mp3 download
This fixes #62 - adding mp3 support.  The major changes to clAudio.py
are the addition of new parameters to ffmpeg.sh -  so it can encode the
mp3 tags.  I also re-tabbed it (was all spaces) so that's why most of
the changes.

clFunctions attempts to put a group value into the page -which may have
already been there - the problem is with no group in the page when it's
media are being built - may just need to load the group at that point
to get the group title.

ffmpeg.sh now builds the mp3 when the generateMP3 flag is "yes" and
this includes the mp3 tags.

makeAllmp3s should be a one-off but it does allow regeneration of all
the mp3 files.

Finally updated the button CSS so the new buttons will match the
locators.