Theading - Part I
This implements the first and arguably harder part of #65 0 moves the
frame generation code out of imagemaker.make_images and into a
"make_image" method of the new Frame_maker class.   It's working and it
should be a fairly straight forward thing to multi-thread it.

Fixed the TDAT (date) tag.