Threading, Part 2 of 2
This one seems to do it - looks like if fixes #65  - finished up on the
threading work, also added a lock to the prog_bar that can be used to
prevent collisions, plus some smarts about out of order frames - a
fairly common occurence now that 8 of them are being built at once.

Turned read delays back on in the ffmpeg output parsig - not sure it's
helping.

Seeing about a 4x speed up in most cases.