The Great Resize
This fixes a bunch of alignment and sizing issues that esp. plagued the
mobile platforms.    Also addresses *some* of what's mentioned in #61,
but the displayed images in the app are all screwed up.   Also -
there's a nagging problem where the base graphic for the Posters is the
wrong resolution (tiny vs. small) and not up to date (combined graphic
still shows split tracks, etc.)

BUT- this does allow the buttons to actually be seen on the iPhone/iPad
and it's surprisingly useful.