Convert-vars and friends
This mostly fixes the problem in convert vars: was doing the
type checking incorrectly.  Also - the group, page, and song
now all set the initial value of createime and updatetme to
a datetime.datetime *not* a string.   That cause a raft of
problems.

Can finally create a page from scratch again.
-jk