Various tweaks, group fix
A number of fixes - probably the most significant, no longer blows
when trying to edit a new group with no songs or pages - but a long
way from being able to create a group from scratch in the tool.
This change was done in coLab.py - and it creates a Label rather than
an OptionMenu if the page or song list is empty.

Turned explicit link generation back on in htmlgen.linkgen.

Added a cute feature to clean up the "short" dates - any date over
180 days the time of day is replaced by the year.   Also found the
%-d (and %-I) strftime formats to suppress leading zeros.

Beyond that a couple of small bug fixes, some comments on what needs to
be done to add a new group in htmlgen.py, and a bit of vertical spacing
for the comment block.