Fixes clicks and keypress
This commit fixes #78 - clicks now do pausing differently if the left
button is held (pause at the play location of the click) or double
clicked (pause at the click location).  They keypress events were fixed
by or’ing with event.which - not sure why that fixed it - since it was
working fine before.