Partial Fix to #80  "stopping"
Doesn’t really stop yet - but a start (ha hah).  The initial image
generation terminates, but the image seq. needs to have it’s threads
stop and need to find a way to kill ffmpeg.

Also need to reset the page.stop flag once everything completes - and
get rid of the window - if that doesn’t already happen