Most of the stopping #80
This implements most of the stopping functions - a bit ugly in ffmpeg
as I was not able to get a useable PID out of the shell - another
argument for scheduling ffmpeg directly.

One known issue (at least): when killing the render - the current
window is left behind.

May want to reset the page.stop to False when we’re done - depends also
on if we want all subsequent runs to be stopped for the given page -
probably we do.